# contracts/openapi/core.v1.yaml
openapi: 3.1.0
info:
  title: ESWMP Core API
  version: v1
  description: |
    Scheduling primitives: Resources, Calendars, Availability Rules, Reservations,
    and Appointments. This service owns scheduling and nothing else — it never
    stores or returns any data belonging to the caller's own domain beyond the
    opaque externalReferenceType/externalReferenceId pair. See docs/ESWMP_VISION.md
    "Key rule".
servers:
  - url: http://localhost:6001
    description: Local development
security:
  - BearerAuth: []
paths:
...
  /api/v1/capacity/profiles:
    post:
      tags: [Capacity]
      summary: Create a Capacity profile for a Resource
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCapacityProfileRequest' }
      responses:
        '201': { description: Profile created, content: { application/json: { schema: { $ref: '#/components/schemas/CapacityProfile' } } } }
  /api/v1/capacity/profiles/{id}:
    get:
      tags: [Capacity]
      summary: Get a Capacity profile
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Profile detail, content: { application/json: { schema: { $ref: '#/components/schemas/CapacityProfile' } } } }
        '404': { description: Not found }
  /api/v1/capacity/profiles/{id}/definitions:
    post:
      tags: [Capacity]
      summary: Add a Capacity definition (one measurable dimension) to a profile
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCapacityDefinitionRequest' }
      responses:
        '201': { description: Definition created, content: { application/json: { schema: { $ref: '#/components/schemas/CapacityDefinition' } } } }
        '404': { description: Profile not found }
  /api/v1/capacity/definitions/{id}:
    patch:
      tags: [Capacity]
      summary: Update a Capacity definition
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateCapacityDefinitionRequest' }
      responses:
        '200': { description: Definition updated, content: { application/json: { schema: { $ref: '#/components/schemas/CapacityDefinition' } } } }
        '404': { description: Not found }
  /api/v1/capacity/resolve:
    post:
      tags: [Capacity]
      summary: Compute effective/held/consumed/remaining capacity for a Resource dimension over a window
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResolveCapacityRequest' }
      responses:
        '200': { description: Resolved capacity, content: { application/json: { schema: { $ref: '#/components/schemas/ResolveCapacityResponse' } } } }
        '404': { description: No active definition for the resource/dimension }
  /api/v1/capacity/explain:
    post:
      tags: [Capacity]
      summary: Breakdown of how a capacity resolution was computed (defined/effective/held/consumed/remaining)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResolveCapacityRequest' }
      responses:
        '200': { description: Capacity explanation, content: { application/json: { schema: { $ref: '#/components/schemas/ExplainCapacityResponse' } } } }
        '404': { description: No active definition for the resource/dimension }
  /api/v1/capacity/holds:
    post:
      tags: [Capacity]
      summary: Atomically acquire a capacity hold (fails with 409 if it would exceed remaining capacity)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCapacityHoldRequest' }
      responses:
        '201': { description: Hold acquired, content: { application/json: { schema: { $ref: '#/components/schemas/CapacityHold' } } } }
        '200': { description: Duplicate request replayed by IdempotencyKey — original hold returned unchanged }
        '404': { description: No active definition for the resource/dimension }
        '409': { description: Requested quantity exceeds remaining capacity }
  /api/v1/capacity/holds/{id}:
    get:
      tags: [Capacity]
      summary: Get a Capacity hold
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Hold detail, content: { application/json: { schema: { $ref: '#/components/schemas/CapacityHold' } } } }
        '404': { description: Not found }
  /api/v1/capacity/holds/{id}/commit:
    post:
      tags: [Capacity]
      summary: Convert an active Hold into a durable Consumption record
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Hold committed, content: { application/json: { schema: { $ref: '#/components/schemas/CapacityConsumption' } } } }
        '404': { description: Not found }
        '409': { description: Hold is not Active }
  /api/v1/capacity/holds/{id}/release:
    post:
      tags: [Capacity]
      summary: Release an active Hold without consuming capacity
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Hold released, content: { application/json: { schema: { $ref: '#/components/schemas/CapacityHold' } } } }
        '404': { description: Not found }
        '409': { description: Hold is not Active }
  /api/v1/capacity/consumptions/{id}/release:
    post:
      tags: [Capacity]
      summary: Release a committed Consumption, freeing its capacity back
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Consumption released, content: { application/json: { schema: { $ref: '#/components/schemas/CapacityConsumption' } } } }
        '404': { description: Not found }
        '409': { description: Consumption is not Committed }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateCapacityProfileRequest:
      type: object
      required: [resourceId, name, timezone]
      properties:
        resourceId: { type: string, format: uuid }
        name: { type: string }
        timezone: { type: string }
    CapacityProfile:
      type: object
      required: [id, resourceId, name, timezone, status, version]
      properties:
        id: { type: string, format: uuid }
        resourceId: { type: string, format: uuid }
        name: { type: string }
        timezone: { type: string }
        status: { type: string, enum: [Draft, Active, Suspended, Retired] }
        version: { type: integer }
    CreateCapacityDefinitionRequest:
      type: object
      required: [name, capacityModel, dimensionCode, maximumQuantity, unit, timeBasis]
      properties:
        name: { type: string }
        capacityModel: { type: string, enum: [Exclusive, Concurrent, Quantity, TimeBucket, DailyQuota] }
        dimensionCode: { type: string, example: "CONCURRENT_WORK" }
        maximumQuantity: { type: integer }
        unit: { type: string, enum: [Count, Kg, Litre, Minute, Hour, CubicMetre] }
        timeBasis: { type: string, enum: [Concurrent, PerBucket, PerDay, PerWeek] }
        effectiveFrom: { type: string, format: date, nullable: true }
        effectiveTo: { type: string, format: date, nullable: true }
    UpdateCapacityDefinitionRequest:
      type: object
      properties:
        name: { type: string, nullable: true }
        maximumQuantity: { type: integer, nullable: true }
        status: { type: string, enum: [Draft, Active, Suspended, Retired], nullable: true }
        effectiveFrom: { type: string, format: date, nullable: true }
        effectiveTo: { type: string, format: date, nullable: true }
    CapacityDefinition:
      type: object
      required: [id, capacityProfileId, name, capacityModel, dimensionCode, maximumQuantity, unit, timeBasis, status, version]
      properties:
        id: { type: string, format: uuid }
        capacityProfileId: { type: string, format: uuid }
        name: { type: string }
        capacityModel: { type: string, enum: [Exclusive, Concurrent, Quantity, TimeBucket, DailyQuota] }
        dimensionCode: { type: string }
        maximumQuantity: { type: integer }
        unit: { type: string, enum: [Count, Kg, Litre, Minute, Hour, CubicMetre] }
        timeBasis: { type: string, enum: [Concurrent, PerBucket, PerDay, PerWeek] }
        effectiveFrom: { type: string, format: date, nullable: true }
        effectiveTo: { type: string, format: date, nullable: true }
        status: { type: string, enum: [Draft, Active, Suspended, Retired] }
        version: { type: integer }
    ResolveCapacityRequest:
      type: object
      required: [resourceId, dimensionCode, startTime, endTime, requiredQuantity]
      properties:
        resourceId: { type: string, format: uuid }
        dimensionCode: { type: string }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        requiredQuantity: { type: integer }
    ResolveCapacityResponse:
      type: object
      required: [effectiveCapacity, consumedCapacity, heldCapacity, remainingCapacity, canFulfil]
      properties:
        effectiveCapacity: { type: integer }
        consumedCapacity: { type: integer }
        heldCapacity: { type: integer }
        remainingCapacity: { type: integer }
        canFulfil: { type: boolean }
    ExplainCapacityResponse:
      type: object
      required: [definedCapacity, effectiveCapacity, activeHolds, confirmedConsumption, remainingCapacity]
      properties:
        definedCapacity: { type: integer }
        effectiveCapacity: { type: integer }
        activeHolds: { type: integer }
        confirmedConsumption: { type: integer }
        remainingCapacity: { type: integer }
    CreateCapacityHoldRequest:
      type: object
      required: [resourceId, dimensionCode, quantity, startTime, endTime, holdDurationSeconds, idempotencyKey]
      properties:
        resourceId: { type: string, format: uuid }
        dimensionCode: { type: string }
        quantity: { type: integer }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        holdDurationSeconds: { type: integer, default: 60 }
        sourceType: { type: string, nullable: true, description: "Opaque caller-domain pointer, never inspected." }
        sourceId: { type: string, nullable: true }
        idempotencyKey: { type: string }
    CapacityHold:
      type: object
      required: [id, capacityDefinitionId, resourceId, dimensionCode, quantity, startTime, endTime, status, expiresAt, idempotencyKey, version]
      properties:
        id: { type: string, format: uuid }
        capacityDefinitionId: { type: string, format: uuid }
        resourceId: { type: string, format: uuid }
        dimensionCode: { type: string }
        quantity: { type: integer }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        status: { type: string, enum: [Active, Committed, Released, Expired, Cancelled] }
        expiresAt: { type: string, format: date-time }
        idempotencyKey: { type: string }
        sourceType: { type: string, nullable: true }
        sourceId: { type: string, nullable: true }
        version: { type: integer }
    CapacityConsumption:
      type: object
      required: [id, capacityDefinitionId, resourceId, dimensionCode, quantity, startTime, endTime, status, version]
      properties:
        id: { type: string, format: uuid }
        capacityDefinitionId: { type: string, format: uuid }
        resourceId: { type: string, format: uuid }
        dimensionCode: { type: string }
        quantity: { type: integer }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        sourceType: { type: string, nullable: true }
        sourceId: { type: string, nullable: true }
        status: { type: string, enum: [Committed, Released] }
        version: { type: integer }
...